correction

+++ lib/features/kpis/domain/entities/kpi.dart
@@ -20,7 +20,7 @@ class Kpi {
     String? objrCantidad;
     String? objrNombreCategoria;
     int? totalRegistro;
-    int? porcentaje;
+    double? porcentaje;
     bool? objrValorDifMes;
     List<ArrayUser>? arrayuserasignacion;
     List<Periodicidad>? peobIdPeriodicidad;
